feat(assetType): integrate feedback
- use format date utility function
- create a new component for asset types actions
- test the new component

[#158491245]
