+++ src/components/AssetTypesComponent.jsx
@@ -6,10 +6,9 @@ import { Container, Header, Table, Button, Pagination } from 'semantic-ui-react'
 import _ from 'lodash';
 
 import TableRowComponent from './TableRowComponent';
+import AssetTypesAction from './AssetTypesAction';
 import SideMenuComponent from '../_components/SideMenuComponent';
 import LoaderComponent from '../components/LoaderComponent';
-import ActionComponent from '../components/ActionComponent';
-import ArtModal from './common/ModalComponent';
 import '../_css/AssetTypesComponent.css';
 
 import { loadAssetTypes } from '../_actions/assetTypes.actions';
@@ -26,16 +25,6 @@ export class AssetTypesComponent extends React.Component {
   handlePaginationChange = (e, { activePage }) => {
     this.setState({ activePage });
     this.props.loadAssetTypes(activePage);
-  }
-
-  formatDate = (assetType, key) => {
-    const date = new Date(assetType[key]);
-    const model = {
-      year: 'numeric',
-      month: 'long',
-      day: '2-digit'
-    };
-    return new Intl.DateTimeFormat('en-GB', model).format(date);
   }
 
   getTotalPages = () => Math.ceil(this.props.assetTypesCount / this.state.limit)
@@ -82,26 +71,7 @@ export class AssetTypesComponent extends React.Component {
                     headings={['id', 'asset_sub_category', 'asset_type']}
                   >
                     <Table.Cell>
-                      <ActionComponent
-                        viewWrapper={element => (
-                          <ArtModal
-                            className="inline"
-                            trigger={element}
-                            modalTitle={assetType.asset_type}
-                          >
-                            {
-                              Object.keys(assetType).map(
-                                (key) => {
-                                  if (key === 'created_at' || key === 'last_modified') {
-                                  return <div className="bottom-20" key={key}>{key}: {this.formatDate(assetType, key)}</div>;
-                                }
-                                  return <div className="bottom-20" key={key}>{key}: {assetType[key]}</div>;
-                                }
-                              )
-                            }
-                          </ArtModal>
-                        )}
-                      />
+                      <AssetTypesAction details={assetType} />
                     </Table.Cell>
                   </TableRowComponent>
                 ))
